+++ example/SampleAdSDK/SampleAdSDKAdapter/SampleAdSDKAdapter.h
@@ -18,6 +18,6 @@ FOUNDATION_EXPORT const unsigned char SampleAdSDKAdapterVersionString[];
 #import <SampleAdSDKAdapter/SampleAdapterDelegate.h>
 #import <SampleAdSDKAdapter/SampleAdapterConstants.h>
 #import <SampleAdSDKAdapter/SampleAdapterProtocol.h>
+#import <SampleAdSDKAdapter/SampleAdapterMediatedNativeAd.h>
 #import <SampleAdSDKAdapter/SampleAdapterMediatedNativeAppInstallAd.h>
 #import <SampleAdSDKAdapter/SampleAdapterMediatedNativeContentAd.h>
-