+++ adapters/Nend/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Nend Adapter for Google Mobile Ads SDK for iOS
 
+## Version 4.0.3.0
+- Verified compatibility with nend SDK 4.0.3.
+
 ## Version 4.0.2.0
 - First release in Google Mobile Ads mediation open source project.
 - Added support for reward-based video ads.